html/browsers/browsing-the-web/navigating-across-documents/initial-empty-document/iframe-src-aboutblank-navigate-immediately.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>Navigations immediately after appending iframe with src='about:blank'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<body></body>
<script>
/*
  When an iframe is created it will contain the initial empty document. However,
  when the iframe's src is set to "about:blank", it will trigger a navigation to
  a non-initial about:blank document, which will replace the initial empty
  document.
  These tests verify the behavior of navigations that happem immediately after
  the iframe is created, when the non-initial about:blank navigation just
  started. Since the non-initial about:blank navigation shouldn't have
  committed yet, the iframe should still be on the initial empty document, so
  navigations that start at that timing should do replacement.
*/
"use strict";
const url1 = "/common/blank.html?1";

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank. This would trigger a
  // navigation to a non-initial about:blank document.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Trigger a navigation to url1 through the iframe's src attribute.
  // Since the previous navigation runs asynchronously, the iframe should still
  // be on the initial empty document, and this should do replacement.
  iframe.src = url1;
  // Wait for the latest navigation to finish.
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");
}, "Navigating to a different document with src");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through setting
  // location.href. Since the previous navigation runs asynchronously, the
  // iframe should still be on the initial empty document, and this should do
  // replacement.
  iframe.contentWindow.location.href = url1;
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");
}, "Navigating to a different document with location.href");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through location.assign().
  // Since the previous navigation runs asynchronously, the iframe should still
  // be on the initial empty document, and this should do replacement.
  iframe.contentWindow.location.assign(url1);
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");
}, "Navigating to a different document with location.assign");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through window.open().
  // Since the previous navigation runs asynchronously, the iframe should still
  // be on the initial empty document, and this should do replacement.
  iframe.contentWindow.open(url1, "_self");
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");
}, "Navigating to a different document with window.open");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through clicking an <a>
  // element. Since the previous navigation runs asynchronously, the iframe
  // should still be on the initial empty document, and this should do
  // replacement.
  const a = iframe.contentDocument.createElement("a");
  a.href = url1;
  iframe.contentDocument.body.appendChild(a);
  a.click();
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");
}, "Navigating to a different document with link click");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through form submission.
  // Since the previous navigation runs asynchronously, the iframe should still
  // be on the initial empty document, and this should do replacement.
  const form = iframe.contentDocument.createElement("form");
  form.action = "/common/blank.html";
  iframe.contentDocument.body.appendChild(form);
  const input = iframe.contentDocument.createElement("input");
  input.type = "hidden";
  input.name = "1";
  form.append(input);
  form.submit();
  await waitForLoad(t, iframe, url1 + "=");
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");
}, "Navigating to a different document with form submission");
</script>
